refactor: fix test structure and names
I noticed that the tests for per-client and per-allocator RSRs are
incorrectly nested inside a `describe()` block for a different feature.

I also noticed one test name refers to "aggregators" instead of
"allocators".

This commit is fixing both problems.

The change is best reviewed in a diff ignoring white-space changes.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>