fix: exclude IPNI 5xx from "success rate" in Influx
In the previous change, I modified the code writing RSR to the database
for spark-stats to exclude IPNI 5xx results from the calculation.

We need to apply the same change to RSR written to InfluxDB.
- We want to RSR in our internal dashboard to be consistent with the RSR
  in the public dashboard.
- I want to create an alert to let us know when the RSR unexpectedly
  drops. That alert must not be triggered by IPNI service degradations.

Note: I still want the result breakdowns to include IPNI 5xx errors.
I added a test to capture that requirement.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>